feat: navbar displays logged in user and can logout
the navbar will display the username belonging to the current active session and add a button to sign out

Refs: #26

+++ labman/src/components/core/Button.tsx
@@ -1,23 +1,40 @@
 ﻿"use client"
-import { deleteUser } from "@/lib/actions";
+import { deleteUser, logout } from "@/lib/actions";
+import {redirect} from "next/navigation";
 
 interface ButtonProps {
+    type?: string;
     username?: string;
 }
 
-export default function Button({ username }: ButtonProps) {
+export default function Button({ username, type }: ButtonProps) {
 
     async function deletion(){
         console.log("Button clicked");
         if (username) {
-            await deleteUser(username)
+            if (type == "deleteUser") {
+                await deleteUser(username)
+            }
+
+        } else if (type == "logout") {
+            console.log("Logout clicked");
+            await logout()
+            redirect("/login");
         }
     }
     
+    if (type == "deleteUser") {
+        return (
+            <button onClick={deletion} className="bg-red-600 flex justify-center w-fit px-2 py-0.5 rounded-md">
+                <p className="text-[1.3rem] text-black font-bold">Delete</p>
+            </button>
+        )
+    } else if (type == "logout") {
+        return (
+            <button onClick={deletion} className="bg-red-600 flex justify-center w-fit px-2 py-0.5 rounded-md">
+                <p className="text-[1.3rem] pb-0 text-black font-bold">Logout</p>
+            </button>
+        )
+    }
 
-    return (
-        <button onClick={deletion} className="bg-red-600 flex justify-center w-fit px-2 py-0.5 rounded-md">
-            <p className="text-[1.3rem] text-black font-bold">Delete</p>
-        </button>
-    )
 }
